+++ server/src/module/mind/setting/entities/mind-system-setting.entity.ts
@@ -1,7 +1,8 @@
 import { Column, Entity, PrimaryColumn } from 'typeorm';
+import { BaseEntity } from '../../../../common/entities/base';
 
 @Entity('t_system_setting', { comment: '系统配置表' })
-export class MindSystemSettingEntity {
+export class MindSystemSettingEntity extends BaseEntity {
   @PrimaryColumn({ type: 'varchar', length: 50, name: 'id', comment: '主键id' })
   id: string;
 
@@ -13,8 +14,5 @@ export class MindSystemSettingEntity {
 
   @Column({ type: 'json', name: 'content', nullable: true, comment: '配置项内容' })
   content: any;
-
-  @Column({ type: 'datetime', name: 'create_time', nullable: true, comment: '创建时间' })
-  createTime: Date | null;
 }